backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
alembic==1.13.3
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
